requirements.txt
fastapi>=0.95.0
langchain
langchain-community
langchain-openai
pypdf
openai
tiktoken
faiss-cpu
matplotlib
uvicorn
python-dotenv
